+++ src/main/java/com/castellanos/fuzzylogicgp/logic/ALogic.java
@@ -5,6 +5,7 @@
  */
 package com.castellanos.fuzzylogicgp.logic;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -12,11 +13,11 @@ import java.util.List;
  * @author hp
  */
 public interface ALogic {
-    public Double not(Double v1);
-    public Double imp(Double v1, Double v2);
-    public Double eqv(Double v1, Double v2);
-    public Double and(Double v1, Double v2);
-    public Double or(Double v1, Double v2);
-    public Double forAll(List<Double> values);
-    public Double exist(List<Double> values);
+    public BigDecimal not(BigDecimal v1);
+    public BigDecimal imp(BigDecimal v1, BigDecimal v2);
+    public BigDecimal eqv(BigDecimal v1, BigDecimal v2);
+    public BigDecimal and(BigDecimal v1, BigDecimal v2);
+    public BigDecimal or(BigDecimal v1, BigDecimal v2);
+    public BigDecimal forAll(List<BigDecimal> values);
+    public BigDecimal exist(List<BigDecimal> values);
 }
